requirements.txt
Django==2.2.28
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.8.3
Markdown==3.3.7
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
zipp==3.6.0

